refactor: rename monitor.Listen() to monitor.Containers()
