+++ main.go
@@ -7,7 +7,11 @@ var MONITOR_INTERVAL = 5 * time.Minute
 func main() {
 	monitor := NewMonitor()
 
+	go monitor.Containers()
 	go monitor.Disk()
 	go monitor.Dmesg()
-	monitor.Listen()
+
+	for {
+		time.Sleep(60 * time.Second)
+	}
 }
